Created Selectors.js and changed type of action in EmployeeActions.js
Moved logic for selectors from InformationHeader and TopHeader files to Selectors.js
Changed action type in EmployeeActions.js file when there is error receiving api data
