+++ src/main/frontend/src/store/actions/EmployeesActions.js
@@ -1,16 +1,18 @@
 export const REQUEST_API_DATA = 'REQUEST_API_DATA';
 export const RECEIVE_API_DATA = 'RECEIVE_API_DATA';
+export const RECEIVE_API_DATA_ERROR = 'RECEIVE_API_DATA_ERROR';
 
 export const errorReceivingApiData = error => ({
   payload: {
     error
   },
-  type: REQUEST_API_DATA
+  type: RECEIVE_API_DATA_ERROR
 });
 
 export const requestApiData = () => ({
   type: REQUEST_API_DATA
 });
+
 export const receiveApiData = employeeInformation => ({
   type: RECEIVE_API_DATA,
   employeeInformation
